new updation
new updation

+++ src/lib/encdec_paytm.php
@@ -1,21 +1,24 @@
 <?php
-function encrypt_e($input, $ky){
+function encrypt_e($input, $ky)
+{
 	$ky   = html_entity_decode($ky);
 	$iv = "@@@@&&&&####$$$$";
-	$data = openssl_encrypt ( $input , "AES-128-CBC" , $ky, 0, $iv );
+	$data = openssl_encrypt($input, "AES-128-CBC", $ky, 0, $iv);
 	return $data;
 }
 
-function decrypt_e($crypt, $ky){
+function decrypt_e($crypt, $ky)
+{
 	$ky   = html_entity_decode($ky);
 	$iv = "@@@@&&&&####$$$$";
-	$data = openssl_decrypt ( $crypt , "AES-128-CBC" , $ky, 0, $iv );
+	$data = openssl_decrypt($crypt, "AES-128-CBC", $ky, 0, $iv);
 	return $data;
 }
 
-function generateSalt_e($length) {
+function generateSalt_e($length)
+{
 	$random = "";
-	srand((double) microtime() * 1000000);
+	srand((float) microtime() * 1000000);
 
 	$data = "AbcDE123IJKLMN67QRSTUVWXYZ";
 	$data .= "aBCdefghijklmn123opq45rs67tuv89wxyz";
@@ -28,13 +31,15 @@ function generateSalt_e($length) {
 	return $random;
 }
 
-function checkString_e($value) {
+function checkString_e($value)
+{
 	if ($value == 'null')
 		$value = '';
 	return $value;
 }
 
-function getChecksumFromArray($arrayList, $key, $sort=1) {
+function getChecksumFromArray($arrayList, $key, $sort = 1)
+{
 	if ($sort != 0) {
 		ksort($arrayList);
 	}
@@ -46,8 +51,9 @@ function getChecksumFromArray($arrayList, $key, $sort=1) {
 	$checksum = encrypt_e($hashString, $key);
 	return $checksum;
 }
-function getChecksumFromString($str, $key) {
-	
+function getChecksumFromString($str, $key)
+{
+
 	$salt = generateSalt_e(4);
 	$finalString = $str . "|" . $salt;
 	$hash = hash("sha256", $finalString);
@@ -56,7 +62,8 @@ function getChecksumFromString($str, $key) {
 	return $checksum;
 }
 
-function verifychecksum_e($arrayList, $key, $checksumvalue) {
+function verifychecksum_e($arrayList, $key, $checksumvalue)
+{
 	$arrayList = removeCheckSumParam($arrayList);
 	ksort($arrayList);
 	$str = getArray2StrForVerify($arrayList);
@@ -77,7 +84,8 @@ function verifychecksum_e($arrayList, $key, $checksumvalue) {
 	return $validFlag;
 }
 
-function verifychecksum_eFromStr($str, $key, $checksumvalue) {
+function verifychecksum_eFromStr($str, $key, $checksumvalue)
+{
 	$paytm_hash = decrypt_e($checksumvalue, $key);
 	$salt = substr($paytm_hash, -4);
 
@@ -95,19 +103,19 @@ function verifychecksum_eFromStr($str, $key, $checksumvalue) {
 	return $validFlag;
 }
 
-function getArray2Str($arrayList) {
+function getArray2Str($arrayList)
+{
 	$findme   = 'REFUND';
 	$findmepipe = '|';
 	$paramStr = "";
-	$flag = 1;	
+	$flag = 1;
 	foreach ($arrayList as $key => $value) {
 		$pos = strpos($value, $findme);
 		$pospipe = strpos($value, $findmepipe);
-		if ($pos !== false || $pospipe !== false) 
-		{
+		if ($pos !== false || $pospipe !== false) {
 			continue;
 		}
-		
+
 		if ($flag) {
 			$paramStr .= checkString_e($value);
 			$flag = 0;
@@ -118,7 +126,8 @@ function getArray2Str($arrayList) {
 	return $paramStr;
 }
 
-function getArray2StrForVerify($arrayList) {
+function getArray2StrForVerify($arrayList)
+{
 	$paramStr = "";
 	$flag = 1;
 	foreach ($arrayList as $key => $value) {
@@ -132,72 +141,88 @@ function getArray2StrForVerify($arrayList) {
 	return $paramStr;
 }
 
-function redirect2PG($paramList, $key) {
+function redirect2PG($paramList, $key)
+{
 	$hashString = getchecksumFromArray($paramList);
 	$checksum = encrypt_e($hashString, $key);
 }
 
-function removeCheckSumParam($arrayList) {
+function removeCheckSumParam($arrayList)
+{
 	if (isset($arrayList["CHECKSUMHASH"])) {
 		unset($arrayList["CHECKSUMHASH"]);
 	}
 	return $arrayList;
 }
 
-function getTxnStatus($requestParamList) {
+function getTxnStatus($requestParamList)
+{
 	return callAPI(PAYTM_STATUS_QUERY_URL, $requestParamList);
 }
 
-function getTxnStatusNew($requestParamList) {
+function getTxnStatusNew($requestParamList)
+{
 	return callNewAPI(PAYTM_STATUS_QUERY_NEW_URL, $requestParamList);
 }
 
-function initiateTxnRefund($requestParamList) {
-	$CHECKSUM = getChecksumFromArray($requestParamList,PAYTM_MERCHANT_KEY,0);
+function initiateTxnRefund($requestParamList)
+{
+	$CHECKSUM = getChecksumFromArray($requestParamList, PAYTM_MERCHANT_KEY, 0);
 	$requestParamList["CHECKSUM"] = $CHECKSUM;
 	return callAPI(PAYTM_REFUND_URL, $requestParamList);
 }
 
-function callAPI($apiURL, $requestParamList) {
+function callAPI($apiURL, $requestParamList)
+{
 	$jsonResponse = "";
 	$responseParamList = array();
-	$JsonData =json_encode($requestParamList);
-	$postData = 'JsonData='.urlencode($JsonData);
+	$JsonData = json_encode($requestParamList);
+	$postData = 'JsonData=' . urlencode($JsonData);
 	$ch = curl_init($apiURL);
-	curl_setopt($ch, CURLOPT_CUSTOMREQUEST, "POST");                                                                     
-	curl_setopt($ch, CURLOPT_POSTFIELDS, $postData);                                                                  
-	curl_setopt($ch, CURLOPT_RETURNTRANSFER, true); 
-	curl_setopt ($ch, CURLOPT_SSL_VERIFYHOST, 0);
-	curl_setopt ($ch, CURLOPT_SSL_VERIFYPEER, 0);
-	curl_setopt($ch, CURLOPT_HTTPHEADER, array(                                                                         
-	'Content-Type: application/json', 
-	'Content-Length: ' . strlen($postData))                                                                       
-	);  
-	$jsonResponse = curl_exec($ch);   
-	$responseParamList = json_decode($jsonResponse,true);
+	curl_setopt($ch, CURLOPT_CUSTOMREQUEST, "POST");
+	curl_setopt($ch, CURLOPT_POSTFIELDS, $postData);
+	curl_setopt($ch, CURLOPT_RETURNTRANSFER, true);
+	curl_setopt($ch, CURLOPT_SSL_VERIFYHOST, 0);
+	curl_setopt($ch, CURLOPT_SSL_VERIFYPEER, 0);
+	curl_setopt(
+		$ch,
+		CURLOPT_HTTPHEADER,
+		array(
+			'Content-Type: application/json',
+			'Content-Length: ' . strlen($postData)
+		)
+	);
+	$jsonResponse = curl_exec($ch);
+	$responseParamList = json_decode($jsonResponse, true);
 	return $responseParamList;
 }
 
-function callNewAPI($apiURL, $requestParamList) {
+function callNewAPI($apiURL, $requestParamList)
+{
 	$jsonResponse = "";
 	$responseParamList = array();
-	$JsonData =json_encode($requestParamList);
-	$postData = 'JsonData='.urlencode($JsonData);
+	$JsonData = json_encode($requestParamList);
+	$postData = 'JsonData=' . urlencode($JsonData);
 	$ch = curl_init($apiURL);
-	curl_setopt($ch, CURLOPT_CUSTOMREQUEST, "POST");                                                                     
-	curl_setopt($ch, CURLOPT_POSTFIELDS, $postData);                                                                  
-	curl_setopt($ch, CURLOPT_RETURNTRANSFER, true); 
-	curl_setopt ($ch, CURLOPT_SSL_VERIFYHOST, 0);
-	curl_setopt ($ch, CURLOPT_SSL_VERIFYPEER, 0);
-	curl_setopt($ch, CURLOPT_HTTPHEADER, array(                                                                         
-	'Content-Type: application/json', 
-	'Content-Length: ' . strlen($postData))                                                                       
-	);  
-	$jsonResponse = curl_exec($ch);   
-	$responseParamList = json_decode($jsonResponse,true);
+	curl_setopt($ch, CURLOPT_CUSTOMREQUEST, "POST");
+	curl_setopt($ch, CURLOPT_POSTFIELDS, $postData);
+	curl_setopt($ch, CURLOPT_RETURNTRANSFER, true);
+	curl_setopt($ch, CURLOPT_SSL_VERIFYHOST, 0);
+	curl_setopt($ch, CURLOPT_SSL_VERIFYPEER, 0);
+	curl_setopt(
+		$ch,
+		CURLOPT_HTTPHEADER,
+		array(
+			'Content-Type: application/json',
+			'Content-Length: ' . strlen($postData)
+		)
+	);
+	$jsonResponse = curl_exec($ch);
+	$responseParamList = json_decode($jsonResponse, true);
 	return $responseParamList;
 }
-function getRefundChecksumFromArray($arrayList, $key, $sort=1) {
+function getRefundChecksumFromArray($arrayList, $key, $sort = 1)
+{
 	if ($sort != 0) {
 		ksort($arrayList);
 	}
@@ -209,17 +234,17 @@ function getRefundChecksumFromArray($arrayList, $key, $sort=1) {
 	$checksum = encrypt_e($hashString, $key);
 	return $checksum;
 }
-function getRefundArray2Str($arrayList) {	
+function getRefundArray2Str($arrayList)
+{
 	$findmepipe = '|';
 	$paramStr = "";
-	$flag = 1;	
-	foreach ($arrayList as $key => $value) {		
+	$flag = 1;
+	foreach ($arrayList as $key => $value) {
 		$pospipe = strpos($value, $findmepipe);
-		if ($pospipe !== false) 
-		{
+		if ($pospipe !== false) {
 			continue;
 		}
-		
+
 		if ($flag) {
 			$paramStr .= checkString_e($value);
 			$flag = 0;
@@ -229,22 +254,23 @@ function getRefundArray2Str($arrayList) {
 	}
 	return $paramStr;
 }
-function callRefundAPI($refundApiURL, $requestParamList) {
+function callRefundAPI($refundApiURL, $requestParamList)
+{
 	$jsonResponse = "";
 	$responseParamList = array();
-	$JsonData =json_encode($requestParamList);
-	$postData = 'JsonData='.urlencode($JsonData);
-	$ch = curl_init($apiURL);	
-	curl_setopt ($ch, CURLOPT_SSL_VERIFYHOST, 0);
-	curl_setopt ($ch, CURLOPT_SSL_VERIFYPEER, 0);
+	$JsonData = json_encode($requestParamList);
+	$postData = 'JsonData=' . urlencode($JsonData);
+	$ch = curl_init($apiURL);
+	curl_setopt($ch, CURLOPT_SSL_VERIFYHOST, 0);
+	curl_setopt($ch, CURLOPT_SSL_VERIFYPEER, 0);
 	curl_setopt($ch, CURLOPT_URL, $refundApiURL);
 	curl_setopt($ch, CURLOPT_POST, true);
-	curl_setopt($ch, CURLOPT_POSTFIELDS, $postData);  
-	curl_setopt($ch, CURLOPT_RETURNTRANSFER, true); 
+	curl_setopt($ch, CURLOPT_POSTFIELDS, $postData);
+	curl_setopt($ch, CURLOPT_RETURNTRANSFER, true);
 	$headers = array();
 	$headers[] = 'Content-Type: application/json';
-	curl_setopt($ch, CURLOPT_HTTPHEADER, $headers);  
-	$jsonResponse = curl_exec($ch);   
-	$responseParamList = json_decode($jsonResponse,true);
+	curl_setopt($ch, CURLOPT_HTTPHEADER, $headers);
+	$jsonResponse = curl_exec($ch);
+	$responseParamList = json_decode($jsonResponse, true);
 	return $responseParamList;
 }
